archive/pending-deletion/2026-02-13/backend-mcp-server-python/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic>=1.7.4,<2.0.0,!=1.8,!=1.8.1
requests==2.31.0
python-dotenv==1.0.0
